+++ common/src/io/gmap_reader.h
@@ -31,7 +31,7 @@
 class gmap_reader {
 public:
 	//DATA
-	std::vector < int > pos_bp;
+	std::vector < long int > pos_bp;
 	std::vector < double > pos_cm;
 
 	//CONSTRUCTOR/DESTRUCTOR
